+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2.0.1 ##
 
+* Added column types to getTables() method
 * Added parameter `forceQueryMode` to use for specifying the type of query
 * Execution of scan or scheme query inside active transaction will raise exception